#313 moving avro logical types conversion config to avro serializers
