+++ src/main/java/net/coru/kloadgen/sampler/KafkaConsumerSampler.java
@@ -19,9 +19,6 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.coru.kloadgen.util.ProducerKeysHelper;
-import org.apache.avro.Conversions;
-import org.apache.avro.data.TimeConversions;
-import org.apache.avro.generic.GenericData;
 import org.apache.jmeter.config.Arguments;
 import org.apache.jmeter.protocol.java.sampler.AbstractJavaSamplerClient;
 import org.apache.jmeter.protocol.java.sampler.JavaSamplerContext;
@@ -45,7 +42,6 @@ public class KafkaConsumerSampler extends AbstractJavaSamplerClient implements S
     final var props = properties(context);
     final var topic = context.getParameter(ProducerKeysHelper.KAFKA_TOPIC_CONFIG);
     consumer = new KafkaConsumer<>(props);
-    configGenericData();
 
     consumer.subscribe(Collections.singletonList(topic));
   }
@@ -58,20 +54,6 @@ public class KafkaConsumerSampler extends AbstractJavaSamplerClient implements S
     timeout = Long.parseLong(props.getProperty("timeout.millis"));
     log.debug("Populated properties: {}", props);
     return props;
-  }
-
-  private void configGenericData() {
-    final var genericData = GenericData.get();
-
-    genericData.addLogicalTypeConversion(new TimeConversions.DateConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.LocalTimestampMicrosConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.LocalTimestampMillisConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.TimeMicrosConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.TimeMillisConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.TimestampMicrosConversion());
-    genericData.addLogicalTypeConversion(new TimeConversions.TimestampMillisConversion());
-    genericData.addLogicalTypeConversion(new Conversions.DecimalConversion());
-    genericData.addLogicalTypeConversion(new Conversions.UUIDConversion());
   }
 
   @Override
